Moving all files that can cause case sensitivity problems
See https://github.com/JeffDownie/camda/issues/4
